New Plans!!!
Beside of loving doria, Ofcoz I need to add more functions on the home page. Seems its simple to add this. For slide show,  I can make the image to a iconimage, which the let user press theimage and. launch the url launcher, which launch the a webpage.
TODO:  Makeapi to get "update JSON" ? which includes the bookstore. link and the image, or just the isb then it self load? Both seems ok wor
I prefer the ISBN one i guess.