I Love U!!!!!!!!!
This part is so done!!!!!!
Scanner is working, which will help user input ISBN, the book detail also worked for user to check the ISBN book is valid, the image picker and camera working for picking an image to upload, and after upload, its on. the list. Im so happy today and I hope u can see this lmao, but I guess u wont QAQ.